Moves FeedImageMapper into its own file and sets appropriate access control
