+++ FeedAPIChallenge/Feed API/RemoteFeedLoader.swift
@@ -30,42 +30,3 @@ public final class RemoteFeedLoader: FeedLoader {
 		}
 	}
 }
-
-internal final class FeedImageMapper {
-	private struct Root: Decodable {
-		let items: [Item]
-		var feed: [FeedImage] {
-			return items.map { $0.item }
-		}
-	}
-
-	private struct Item: Decodable {
-		let id: UUID
-		let description: String?
-		let location: String?
-		let url: URL
-
-		private enum CodingKeys: String, CodingKey {
-			case id = "image_id"
-			case description = "image_desc"
-			case location = "image_loc"
-			case url = "image_url"
-		}
-
-		var item: FeedImage {
-			return FeedImage(id: id,
-			                 description: description,
-			                 location: location,
-			                 url: url)
-		}
-	}
-
-	internal static func map(_ data: Data, response: HTTPURLResponse) -> FeedLoader.Result {
-		guard response.statusCode == 200,
-		      let root = try? JSONDecoder().decode(Root.self, from: data) else {
-			return .failure(RemoteFeedLoader.Error.invalidData)
-		}
-
-		return .success(root.feed)
-	}
-}
